[FEAT] Dirty Check Exam

+++ src/main/java/com/ll/jpa/standard/util/Ut.java
@@ -3,10 +3,10 @@ package com.ll.jpa.standard.util;
 import lombok.SneakyThrows;
 
 public class Ut {
-    public static class thread{
+    public static class thread {
 
         @SneakyThrows
-        public static void sleep(long millis) {
+        public static void sleep(int millis) {
             Thread.sleep(millis);
         }
     }
